+++ CHANGELOG.md
@@ -11,6 +11,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 - Log build result in make target.
+- Live log status context menu to reset the status text to the default connected message.
 
 ### Changed
 - Refactor README and update documentation structure.